Employee Payroll App Setup Also Create A Controller To Demostrate The Various HTTP Methods

+++ src/main/java/com/example/emppayrollapp/DTO/EmpPayrollDTO.java
@@ -18,7 +18,7 @@ public class EmpPayrollDTO {
         this.name = name;
     }
 
-    public double getSalary() {
+    public long getSalary() {
         return salary;
     }
